+++ client/client_test.go
@@ -40,14 +40,14 @@ func TestClient(t *testing.T) {
 	})
 
 	convey.Convey("ListProposals", t, func() {
-		list, err := c.ListProposals(context.Background(), 0, 10, 0, nil, "created", OrderDirectionDesc)
+		list, err := c.ListProposals(context.Background(), 0, 10, 0, nil, nil, "created", OrderDirectionDesc)
 
 		convey.So(err, convey.ShouldBeNil)
 		convey.So(list.GetProposals(), convey.ShouldHaveLength, 10)
 	})
 
 	convey.Convey("ListProposals filtered by spaces", t, func() {
-		list, err := c.ListProposals(context.Background(), 0, 10, 0, []*string{helpers.Ptr("stgdao.eth")}, "created", OrderDirectionDesc)
+		list, err := c.ListProposals(context.Background(), 0, 10, 0, []*string{helpers.Ptr("stgdao.eth")}, nil, "created", OrderDirectionDesc)
 
 		convey.So(err, convey.ShouldBeNil)
 		convey.So(list.GetProposals(), convey.ShouldHaveLength, 10)
@@ -63,7 +63,7 @@ func TestClient(t *testing.T) {
 	})
 
 	convey.Convey("ListVotes", t, func() {
-		list, err := c.ListVotes(context.Background(), "0x108a9e597560c4f249cd8be23acd409059fcd17bb2290d69a550ac2232676e7d", 0, 10, "created", OrderDirectionDesc)
+		list, err := c.ListVotes(context.Background(), []*string{helpers.Ptr("0x108a9e597560c4f249cd8be23acd409059fcd17bb2290d69a550ac2232676e7d")}, 0, 10, "created", OrderDirectionDesc)
 
 		convey.So(err, convey.ShouldBeNil)
 		convey.So(list.GetVotes(), convey.ShouldHaveLength, 10)